+++ sentry-android-core/src/main/java/io/sentry/android/core/PerformanceAndroidEventProcessor.java
@@ -24,6 +24,7 @@ import io.sentry.util.Objects;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -248,7 +249,8 @@ final class PerformanceAndroidEventProcessor implements EventProcessor {
         span.getDescription(),
         SpanStatus.OK,
         APP_METRICS_ORIGIN,
-        new HashMap<>(),
+        new ConcurrentHashMap<>(),
+        new ConcurrentHashMap<>(),
         defaultSpanData);
   }
 }